Refine MCP tool telemetry
